Module access server-only. Restructured module setup
New modules must provide a Name. Modules that define a "gamemode" will only be enabled in that gamemode.

Restructured module setup. Module table is setup in init function. Module information is now stored in a global table that is readonly to clients, and read-write on server.

TODO: Fix `_registerDataValue` and `_registerWeaponDataValue` functions to have proper MySQL syntax, not SQL Server.